Particles: Use string as input for selecting distribution function
