--- renpy/display/particle/__init__.py
@@ -1,4 +1,0 @@
-from .particle import Sprite, SpriteManager, Particles, SnowBlossom
-
-
-__all__ = [ 'Sprite', 'SpriteManager', 'Particles', 'SnowBlossom' ]
